Move unify to utils
